MVP FFMPEG installer and 720p encoder.
I have tested this on Ubuntu and Amazon Linux via Docker.

I like the x265 codec and I think 720p is perfectly fine in terms of resolution.

I can not personally tell the difference between x265 720p vs x265 1080p on the TV that I have.

I wanted to learn a bit about ffmpeg considering it is a software which is everywhere.