Add default values and adjust encoding bitrate.
To make the interface easier to use I have gotten rid of the option to
pass the number of threads.

ffmpeg will auto detect the number of threads and use up to 16 of them.

More threads than 16 does not mean you will get better compression or
shorter encode times.

I added an option to be able to pass the encoder preset you want to use.

It default to medium.